When changing the type, remember to change the type.

+++ example/finalize.c
@@ -12,7 +12,7 @@ const int iterations = 100000;
 int live_count = 0;
 
 void finalize(void* data) {
-  int i = (intptr_t)data;
+  intptr_t i = (intptr_t)data;
   if (i % (iterations / 10) == 0) printf("Finalizing #%" PRIdPTR "...\n", i);
   --live_count;
 }
